Improvements to code quality and efficiency.
* Prefer size_t to uint for representing IDs.
* For efficiency, use a vector instead of unordered_map for storing cluster information.
* Minor naming improvements.

+++ r-package/grf/bindings/InstrumentalForestBindings.cpp
@@ -27,8 +27,8 @@ Rcpp::List instrumental_train(Rcpp::NumericMatrix input_data,
                               double reduced_form_weight,
                               double alpha,
                               bool imbalance_penalty,
-                              std::vector<unsigned int> clusters,
-                              unsigned int samples_per_cluster) {
+                              std::vector<size_t> clusters,
+                              uint samples_per_cluster) {
   ForestTrainer trainer = ForestTrainers::instrumental_trainer(
       outcome_index - 1,
       treatment_index - 1,
