+++ r-package/grf/bindings/QuantileForestBindings.cpp
@@ -15,18 +15,18 @@ Rcpp::List quantile_train(std::vector<double> quantiles,
                           Rcpp::NumericMatrix input_data,
                           Eigen::SparseMatrix<double> sparse_input_data,
                           size_t outcome_index,
-                          unsigned int mtry,
-                          unsigned int num_trees,
+                          uint mtry,
+                          uint num_trees,
                           int num_threads,
                           int min_node_size,
                           double sample_fraction,
-                          unsigned int seed,
+                          uint seed,
                           bool honesty,
-                          unsigned int ci_group_size,
+                          uint ci_group_size,
                           double alpha,
                           double imbalance_penalty,
-                          std::vector<unsigned int> clusters,
-                          unsigned int samples_per_cluster) {
+                          std::vector<size_t> clusters,
+                          uint samples_per_cluster) {
   ForestTrainer trainer = regression_splits
       ? ForestTrainers::regression_trainer(outcome_index - 1)
       : ForestTrainers::quantile_trainer(outcome_index - 1, quantiles);
@@ -47,7 +47,7 @@ Rcpp::NumericMatrix quantile_predict(Rcpp::List forest_object,
                                      std::vector<double> quantiles,
                                      Rcpp::NumericMatrix input_data,
                                      Eigen::SparseMatrix<double> sparse_input_data,
-                                     unsigned int num_threads) {
+                                     uint num_threads) {
   Data* data = RcppUtilities::convert_data(input_data, sparse_input_data);
   Forest forest = RcppUtilities::deserialize_forest(
       forest_object[RcppUtilities::SERIALIZED_FOREST_KEY]);
@@ -65,7 +65,7 @@ Rcpp::NumericMatrix quantile_predict_oob(Rcpp::List forest_object,
                                          std::vector<double> quantiles,
                                          Rcpp::NumericMatrix input_data,
                                          Eigen::SparseMatrix<double> sparse_input_data,
-                                         unsigned int num_threads) {
+                                         uint num_threads) {
   Data* data = RcppUtilities::convert_data(input_data, sparse_input_data);
   Forest forest = RcppUtilities::deserialize_forest(
       forest_object[RcppUtilities::SERIALIZED_FOREST_KEY]);
